promote type and allocate if necessary

+++ test/rulesets/LinearAlgebra/factorization.jl
@@ -135,7 +135,7 @@ const LU_NO_PIVOT = VERSION >= v"1.7.0-DEV.1188" ? NoPivot() : Val(false)
             Z = randn(Float32, 10, 10)
             result = ChainRules._mulsubtrans!!(copy(Z), Y)
             @test result ≈ Y .* (Z - Z')
-            @test eltype(result) == Float32
+            @test eltype(result) == Float64
         end
     end
